Moving check_nonzero_size to arr and fixing behavior to act like it did before
