+++ stan/math/fwd/mat/fun/qr_Q.hpp
@@ -1,11 +1,11 @@
 #ifndef STAN_MATH_FWD_MAT_FUN_QR_Q_HPP
 #define STAN_MATH_FWD_MAT_FUN_QR_Q_HPP
 
+#include <stan/math/prim/arr/err/check_nonzero_size.hpp>
 #include <stan/math/prim/mat/fun/Eigen.hpp>
-#include <Eigen/QR>
-#include <stan/math/prim/scal/err/check_nonzero_size.hpp>
 #include <stan/math/prim/scal/err/check_greater_or_equal.hpp>
 #include <stan/math/fwd/core.hpp>
+#include <Eigen/QR>
 
 namespace stan {
   namespace math {
