+++ stan/math/prim/arr.hpp
@@ -11,6 +11,7 @@
 #include <stan/math/prim/arr/meta/VectorBuilderHelper.hpp>
 #include <stan/math/prim/arr/meta/VectorView.hpp>
 
+#include <stan/math/prim/arr/err/check_nonzero_size.hpp>
 #include <stan/math/prim/arr/err/check_ordered.hpp>
 
 #include <stan/math/prim/arr/fun/dist.hpp>
