+++ stan/math/prim/mat/err/check_simplex.hpp
@@ -1,11 +1,11 @@
 #ifndef STAN_MATH_PRIM_MAT_ERR_CHECK_SIMPLEX_HPP
 #define STAN_MATH_PRIM_MAT_ERR_CHECK_SIMPLEX_HPP
 
-#include <stan/math/prim/scal/err/domain_error.hpp>
+#include <stan/math/prim/arr/err/check_nonzero_size.hpp>
 #include <stan/math/prim/mat/err/constraint_tolerance.hpp>
-#include <stan/math/prim/scal/err/check_nonzero_size.hpp>
 #include <stan/math/prim/mat/fun/Eigen.hpp>
 #include <stan/math/prim/mat/meta/index_type.hpp>
+#include <stan/math/prim/scal/err/domain_error.hpp>
 #include <stan/math/prim/scal/meta/error_index.hpp>
 #include <sstream>
 #include <string>
