+++ stan/math/prim/mat/fun/mean.hpp
@@ -1,8 +1,8 @@
 #ifndef STAN_MATH_PRIM_MAT_FUN_MEAN_HPP
 #define STAN_MATH_PRIM_MAT_FUN_MEAN_HPP
 
+#include <stan/math/prim/arr/err/check_nonzero_size.hpp>
 #include <stan/math/prim/mat/fun/Eigen.hpp>
-#include <stan/math/prim/scal/err/check_nonzero_size.hpp>
 #include <boost/math/tools/promotion.hpp>
 #include <vector>
