+++ stan/math/prim/mat/fun/qr_Q.hpp
@@ -1,9 +1,9 @@
 #ifndef STAN_MATH_PRIM_MAT_FUN_QR_Q_HPP
 #define STAN_MATH_PRIM_MAT_FUN_QR_Q_HPP
 
-#include <stan/math/prim/scal/err/check_greater_or_equal.hpp>
-#include <stan/math/prim/scal/err/check_nonzero_size.hpp>
+#include <stan/math/prim/arr/err/check_nonzero_size.hpp>
 #include <stan/math/prim/mat/fun/Eigen.hpp>
+#include <stan/math/prim/scal/err/check_greater_or_equal.hpp>
 #include <Eigen/QR>
 
 namespace stan {
