+++ stan/math/prim/mat/fun/sd.hpp
@@ -1,9 +1,9 @@
 #ifndef STAN_MATH_PRIM_MAT_FUN_SD_HPP
 #define STAN_MATH_PRIM_MAT_FUN_SD_HPP
 
+#include <stan/math/prim/arr/err/check_nonzero_size.hpp>
 #include <stan/math/prim/mat/fun/Eigen.hpp>
 #include <stan/math/prim/mat/fun/variance.hpp>
-#include <stan/math/prim/scal/err/check_nonzero_size.hpp>
 #include <boost/math/tools/promotion.hpp>
 #include <vector>
