+++ stan/math/rev/mat/fun/variance.hpp
@@ -5,7 +5,7 @@
 #include <stan/math/prim/mat/fun/Eigen.hpp>
 #include <stan/math/prim/mat/fun/mean.hpp>
 #include <stan/math/rev/core.hpp>
-#include <stan/math/prim/scal/err/check_nonzero_size.hpp>
+#include <stan/math/prim/arr/err/check_nonzero_size.hpp>
 #include <vector>
 
 namespace stan {
